+++ src/scenes/plain.cpp
@@ -38,9 +38,9 @@ Plain::Plain(): Scene(2, 2, 130.0f) {
         return materialStatic{
           .castsShadows = true,
           .ambient = Color(0.0f, 0.025f, 0.0f),
-          .diffuse = Color((sinf(point.x / 2 + point.y + point.z + frame / 5) + 0.3) * 0.5,
-                           (sinf(point.x / 4 + point.y / 2 + point.z + frame / 10) + 0.4) * 0.7,
-                           (sinf(point.x / 8 + point.y + point.z / 4 + frame / 20) + 0.2) * 0.3),
+          .diffuse = Color((sinf(point.x / 2 + point.y     + point.z     + frame / 5)  + 0.3f) * 0.5f,
+                           (sinf(point.x / 4 + point.y / 2 + point.z     + frame / 10) + 0.4f) * 0.7f,
+                           (sinf(point.x / 8 + point.y     + point.z / 4 + frame / 20) + 0.2f) * 0.3f),
           .specular = Color(0.2f),
           .emission = Color(0.0),
           .shininess = point.z,
@@ -53,7 +53,10 @@ Plain::Plain(): Scene(2, 2, 130.0f) {
 
 
   objects[1] = SphereGen([](float frame) {
-    return Sphere(Vector3(sinf(frame / 8) * 15, sin(frame / 4) * 2 + 9, sin(frame / 20) * 45 + 65), 10,
+    return Sphere(Vector3(sinf(frame / 8) * 15,
+                          sinf(frame / 4) * 2 + 9,
+                          sinf(frame / 20) * 45 + 65),
+                  10,
                   [](Vector3 point, float frame) {
                     return Materials::red;
                   });
